Rewrote the entire codebase to adapt to jpm and no binaries
This was rewritten using jpm Firefox extension sdk to support upscaling
images. Removed all c++ code to resize/compress images and now use a
modified version of pica library with canvas and JavaScript. Logically
created more classes to separate the logic to avoid edge cases. Supports
upscaling for pngs with transparency and auto changes them to jpeg when
there is no transparency if placed as a setting. Changed upscaling
source to Waifu2xBooru because online was too slow and had smaller
scaling parameters and sometimes errored unexpectedly.

Settings do not work and there may be glitches such as images are
skipped to be upscaled.